# openapi-spec.yaml
openapi: 3.0.3
info:
  title: MyLab Platform API
  description: Complete API specification for the MyLab Platform backend
  version: 1.1.0
  contact:
    name: MyLab Platform Team
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.mylab-platform.com/api
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /projects/{id}:
    get:
      summary: Get project
      description: Get project details
      tags: [Projects]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Project'

    put:
      summary: Update project
      description: Update project
      tags: [Projects]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Project'
                  message:
                    type: string

    delete:
      summary: Delete project
      description: Delete project (admin required)
      tags: [Projects]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Project deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # ============================================================================
  # PROJECT STAGES (Nested)
  # ============================================================================
...
  /notifications:
    get:
      summary: List notifications
      description: Get a paginated list of notifications for the current user
      tags: [Notifications]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [unread, read, archived]
        - name: type
          in: query
          schema:
            type: string
            enum: [info, warning, error, success]
      responses:
        '200':
          description: List of notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /notifications/{id}:
    get:
      summary: Get notification by ID
      description: Get detailed information about a specific notification
      tags: [Notifications]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Notification details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Notification not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update notification
      description: Update notification status (mark as read/unread)
      tags: [Notifications]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNotificationRequest'
      responses:
        '200':
          description: Notification updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Notification not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message"
        details:
          type: object
          description: Additional error details
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5

    # ============================================================================
    # AUTH SCHEMAS
    # ============================================================================
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 2000
        clientOrgId:
          type: string
          format: uuid
        clientOrgName:
          type: string
        externalClientName:
          type: string
        executingOrgId:
          type: string
          format: uuid
        executingOrgName:
          type: string
        status:
          type: string
          enum: [active, completed, archived]
          default: active
        workflowMode:
          type: string
          enum: [analysis_first, trial_first]
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
...
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 2000
        status:
          type: string
          enum: [active, completed, archived]
        workflowMode:
          type: string
          enum: [analysis_first, trial_first]

    # ============================================================================
    # STAGE SCHEMAS
    # ============================================================================
...
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [info, warning, error, success]
          default: info
        status:
          type: string
          enum: [unread, read, archived]
          default: unread
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: medium
        category:
          type: string
          description: Notification category (e.g., analysis, project, system)
        metadata:
          type: object
          description: Additional notification data
        actionUrl:
          type: string
          format: uri
          description: URL to navigate to when notification is clicked
        expiresAt:
          type: string
          format: date-time
        readAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateNotificationRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [read, archived]
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...